fix Errors on uploadPhoto
